Add sidebar to dashboard
