+++ packages/js/src/dashboard/components/dashboard.js
@@ -13,12 +13,12 @@ import { PageTitle } from "./page-title";
  */
 export const Dashboard = ( { contentTypes, userName } ) => {
 	return (
-		<div className="yst-@container">
+		<>
 			<PageTitle userName={ userName } />
 			<div className="yst-flex yst-flex-col @7xl:yst-flex-row yst-gap-6 yst-mt-6">
 				<SeoScores contentTypes={ contentTypes } />
 				<ReadabilityScores contentTypes={ contentTypes } />
 			</div>
-		</div>
+		</>
 	);
 };
