+++ packages/js/src/general/routes/alert-center.js
@@ -1,7 +1,7 @@
 import { useSelect } from "@wordpress/data";
 import { __ } from "@wordpress/i18n";
 import { Paper, Title } from "@yoast/ui-library";
-import { PremiumUpsellList, SidebarRecommendations } from "../../shared-admin/components";
+import { PremiumUpsellList } from "../../shared-admin/components";
 import { Notifications, Problems } from "../components";
 import { STORE_NAME } from "../constants";
 import { useSelectGeneralPage } from "../hooks";
@@ -12,13 +12,11 @@ import { useSelectGeneralPage } from "../hooks";
 export const AlertCenter = () => {
 	const isPremium = useSelectGeneralPage( "selectPreference", [], "isPremium" );
 	const premiumLinkList = useSelectGeneralPage( "selectLink", [], "https://yoa.st/17h" );
-	const premiumLinkSidebar = useSelectGeneralPage( "selectLink", [], "https://yoa.st/jj" );
 	const premiumUpsellConfig = useSelectGeneralPage( "selectUpsellSettingsAsProps" );
-	const academyLink = useSelectGeneralPage( "selectLink", [], "https://yoa.st/3t6" );
 	const { isPromotionActive } = useSelect( STORE_NAME );
 
-	return <div className="yst-flex yst-gap-6 xl:yst-flex-row yst-flex-col">
-		<div className="yst-@container yst-flex yst-flex-wrap yst-flex-grow yst-flex-col">
+	return (
+		<>
 			<Paper className="yst-p-8 yst-grow">
 				<header className="yst-max-w-screen-sm">
 					<Title>{ __( "Alert center", "wordpress-seo" ) }</Title>
@@ -31,23 +29,13 @@ export const AlertCenter = () => {
 				<Problems />
 				<Notifications />
 			</div>
-			{ ! isPremium && <PremiumUpsellList
-				premiumLink={ premiumLinkList }
-				premiumUpsellConfig={ premiumUpsellConfig }
-				isPromotionActive={ isPromotionActive }
-			/> }
-		</div>
-		{ ! isPremium &&
-			<div className="yst-min-w-[16rem] xl:yst-max-w-[16rem]">
-				<div className="yst-sticky yst-top-16">
-					<SidebarRecommendations
-						premiumLink={ premiumLinkSidebar }
-						premiumUpsellConfig={ premiumUpsellConfig }
-						academyLink={ academyLink }
-						isPromotionActive={ isPromotionActive }
-					/>
-				</div>
-			</div>
-		}
-	</div>;
+			{ isPremium ? null : (
+				<PremiumUpsellList
+					premiumLink={ premiumLinkList }
+					premiumUpsellConfig={ premiumUpsellConfig }
+					isPromotionActive={ isPromotionActive }
+				/>
+			) }
+		</>
+	);
 };
